revising: IDE: script runner: error handling (ongoing)
