+++ API/src/main/java/org/sikuli/script/runnerHelpers/JythonHelper.java
@@ -10,6 +10,8 @@ import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.sikuli.basics.Debug;
 import org.sikuli.basics.FileManager;
@@ -878,4 +880,175 @@ public class JythonHelper implements IScriptLanguageHelper {
     return trace;
   }
 
+  public int findErrorSource(Throwable thr, String filename) {
+    errorClass = PY_UNKNOWN;
+    String err = "ERROR_UNKNOWN";
+    try {
+      err = thr.toString();
+    } catch (Exception ex) {
+      errorClass = PY_JAVA;
+      err = thr.getCause().toString();
+    }
+//    log(-1,"------------- Traceback -------------\n" + err +
+//            "------------- Traceback -------------\n");
+    errorLine = -1;
+    errorColumn = -1;
+    errorType = "--UnKnown--";
+    errorText = "--UnKnown--";
+
+//  File ".../mainpy.sikuli/mainpy.py", line 25, in <module> NL func() NL
+//  File ".../subpy.py", line 4, in func NL 1/0 NL
+    Pattern pFile = Pattern.compile("File..(.*?\\.py).*?" + ",.*?line.*?(\\d+),.*?in(.*?)" + NL + "(.*?)" + NL);
+
+    String msg;
+    Matcher mFile = null;
+
+    if (PY_JAVA != errorClass) {
+      if (err.startsWith("Traceback")) {
+        Pattern pError = Pattern.compile(NL + "(.*?):.(.*)$");
+        mFile = pFile.matcher(err);
+        if (mFile.find()) {
+          log(lvl + 2, "Runtime error line: " + mFile.group(2) + "\n in function: " + mFile.group(3) + "\n statement: "
+                  + mFile.group(4));
+          errorLine = Integer.parseInt(mFile.group(2));
+          errorClass = PY_RUNTIME;
+          Matcher mError = pError.matcher(err);
+          if (mError.find()) {
+            log(lvl + 2, "Error:" + mError.group(1));
+            log(lvl + 2, "Error:" + mError.group(2));
+            errorType = mError.group(1);
+            errorText = mError.group(2);
+          } else {
+//org.sikuli.core.FindFailed: FindFailed: can not find 1352647716171.png on the screen
+            Pattern pFF = Pattern.compile(": FindFailed: (.*?)" + NL);
+            Matcher mFF = pFF.matcher(err);
+            if (mFF.find()) {
+              errorType = "FindFailed";
+              errorText = mFF.group(1);
+            } else {
+              errorClass = PY_UNKNOWN;
+            }
+          }
+        }
+      } else if (err.startsWith("SyntaxError")) {
+        Pattern pLineS = Pattern.compile(", (\\d+), (\\d+),");
+        java.util.regex.Matcher mLine = pLineS.matcher(err);
+        if (mLine.find()) {
+          log(lvl + 2, "SyntaxError error line: " + mLine.group(1));
+          Pattern pText = Pattern.compile("\\((.*?)\\(");
+          java.util.regex.Matcher mText = pText.matcher(err);
+          mText.find();
+          errorText = mText.group(1) == null ? errorText : mText.group(1);
+          log(lvl + 2, "SyntaxError: " + errorText);
+          errorLine = Integer.parseInt(mLine.group(1));
+          errorColumn = Integer.parseInt(mLine.group(2));
+          errorClass = PY_SYNTAX;
+          errorType = "SyntaxError";
+        }
+      }
+    }
+
+    msg = String.format("script [ %s ]", new File(filename).getName().replace(".py", ""));
+
+    if (errorLine != -1) {
+      // log(-1,_I("msgErrorLine", srcLine));
+      msg += " stopped with error in line " + errorLine;
+      if (errorColumn != -1) {
+        msg += " at column " + errorColumn;
+      }
+    } else {
+      msg += " stopped with error at line --unknown--";
+    }
+
+    if (errorClass == PY_RUNTIME || errorClass == PY_SYNTAX) {
+      if (errorText.startsWith(errorType)) {
+        if (errorText.startsWith("java.lang.RuntimeException: SikuliX: ")) {
+          errorText = errorText.replace("java.lang.RuntimeException: SikuliX: ", "sikulix.RuntimeException: ");
+          errorClass = SX_RUNTIME;
+        } else if (errorText.startsWith("java.lang.ThreadDeath")) {
+          errorClass = SX_ABORT;
+        }
+      }
+      if (errorClass != SX_ABORT) {
+        Debug.error(msg);
+        Debug.error(errorType + " ( " + errorText + " )");
+      } else {
+        Debug.error("IDE: terminating script run - abort key was pressed");
+      }
+      if (errorClass == PY_RUNTIME) {
+        errorTrace = findErrorSourceWalkTrace(mFile, filename);
+        if (errorTrace.length() > 0) {
+          Debug.error("--- Traceback --- error source first\n" + "line: module ( function ) statement \n" + errorTrace
+                  + "[error] --- Traceback --- end --------------");
+        }
+      }
+    } else {
+      Debug.error(msg);
+      Debug.error("Error caused by: %s", err);
+    }
+    return errorLine;
+  }
+
+  private String findErrorSourceWalkTrace(Matcher m, String filename) {
+    Pattern pModule;
+    if (runTime.runningWindows) {
+      pModule = Pattern.compile(".*\\\\(.*?)\\.py");
+    } else {
+      pModule = Pattern.compile(".*/(.*?)\\.py");
+    }
+    String mod;
+    String modIgnore = "SikuliImporter,Region,";
+    StringBuilder trace = new StringBuilder();
+    String telem;
+    while (m.find()) {
+      if (m.group(1).equals(filename)) {
+        mod = "main";
+      } else {
+        Matcher mModule = pModule.matcher(m.group(1));
+        mModule.find();
+        mod = mModule.group(1);
+        if (modIgnore.contains(mod + ",")) {
+          continue;
+        }
+      }
+      telem = m.group(2) + ": " + mod + " ( " + m.group(3) + " ) " + m.group(4) + NL;
+      trace.insert(0, telem);
+    }
+    return trace.toString();
+  }
+
+  private void findErrorSourceFromJavaStackTrace(Throwable thr, String filename) {
+    log(-1, "findErrorSourceFromJavaStackTrace: seems to be an error in the Java API supporting code");
+    StackTraceElement[] s;
+    Throwable t = thr;
+    while (t != null) {
+      s = t.getStackTrace();
+      log(lvl + 2, "stack trace:");
+      for (int i = s.length - 1; i >= 0; i--) {
+        StackTraceElement si = s[i];
+        log(lvl + 2, si.getLineNumber() + " " + si.getFileName());
+        if (si.getLineNumber() >= 0 && filename.equals(si.getFileName())) {
+          errorLine = si.getLineNumber();
+        }
+      }
+      t = t.getCause();
+      log(lvl + 2, "cause: " + t);
+    }
+  }
+
+  private int errorLine;
+  private int errorColumn;
+  private String errorType;
+  private String errorText;
+  private int errorClass;
+  private String errorTrace;
+
+  private static final int PY_SYNTAX = 0;
+  private static final int PY_RUNTIME = 1;
+  private static final int PY_JAVA = 2;
+  private static final int SX_RUNTIME = 3;
+  private static final int SX_ABORT = 4;
+  private static final int PY_UNKNOWN = -1;
+
+  private static final String NL = String.format("%n");
 }
