+++ API/src/main/java/org/sikuli/script/runners/JavaScriptRunner.java
@@ -48,6 +48,7 @@ public class JavaScriptRunner extends AbstractScriptRunner {
 
   @Override
   protected void doInit(String[] args) throws Exception {
+    Debug.info("JavaScript init");
     ScriptEngineManager jsFactory = new ScriptEngineManager();
     ScriptEngine engine = jsFactory.getEngineByName("JavaScript");
